# api/docs/swagger.yaml
definitions:
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateMediaModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchMediaModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateMediaModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  memory.GetAllMediaResponse:
    properties:
      count:
        type: integer
      media:
        items:
          $ref: '#/definitions/memory.Media'
        type: array
    type: object
...
  memory.Media:
    properties:
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/media:
    get:
      consumes:
      - application/json
      description: Get a list of media.
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        minimum: 1
        name: limit
        type: integer
      - description: Filter by memory ID
        in: query
        name: memory_id
        type: string
      - description: Filter by media types (comma-separated)
        in: query
        name: types
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllMediaResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Media
      tags:
      - Media
    post:
      consumes:
      - application/json
      description: Create new media with the provided details.
      parameters:
      - description: Media details
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateMediaModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create new Media
      tags:
      - Media
  /v1/media/{id}:
    delete:
      consumes:
      - application/json
      description: Delete media by its ID.
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Media
      tags:
      - Media
    get:
      consumes:
      - application/json
      description: Get details of media by its ID.
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Media'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Media by ID
      tags:
      - Media
    patch:
      consumes:
      - application/json
      description: Partially update existing media.
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Patched media data
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchMediaModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Partially update Media
      tags:
      - Media
    put:
      consumes:
      - application/json
      description: Update existing media.
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated media data
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateMediaModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Media
      tags:
      - Media
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"